Move `AccessibilePopoverContent` logic to separate hook
