+++ packages/components/src/accessible-popover/accessible-popover-content/index.js
@@ -1,1 +1,2 @@
 export { default } from './component';
+export { useAccessiblePopoverContent } from './hook';
